# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nxs-anomaly native API",
    "version": "1.9.17",
    "description": "The native /api/v1 contract for nxs-anomaly (API, worker, frontend). This is the authoritative contract; the Grafana-compatibility surface under /api/internal/v1 is a compatibility shim and is NOT part of this contract. The core payloads — user, team, schedule, escalation chain, integration, alert group, alert, notification, delivery attempt — are modelled field by field, with `required` reflecting exactly the keys the engine writes unconditionally; every one still allows additional properties because the underlying storage is JSONB. Remaining resources use the open `Entity` schema. Drift between this document and the router is gated by internal/server/openapi_contract_test.go, and the frontend compiles against types generated from this file (frontend/src/api/schema.d.ts).",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin (frontend nginx proxies /api to the API service)"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "apiKeyHeader": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List users",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserList"
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/User"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nxs_anomaly_session",
        "description": "Browser session cookie issued by /api/v1/auth/login."
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key mapped to a role."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Static API key sent as `Authorization: Bearer <key>`, or a mobile session token (prefix `nxm_`). A mobile session acts as its user, with the role capped at responder."
      },
      "mobileSession": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Mobile-Session",
        "description": "Mobile session token issued by /api/v1/mobile/pairing/redeem (or, administratively, /api/v1/mobile/sessions)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "description": "The single error envelope used by every failure response."
      },
      "Entity": {
        "type": "object",
        "description": "A domain object stored as a JSON document, for resources this contract does not model field by field. Always carries an `id`; other fields depend on the resource.",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "NotificationTarget": {
        "type": "object",
        "required": [
          "type",
          "target"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "log",
              "webhook",
              "telegram",
              "email",
              "call",
              "slack",
              "mattermost"
            ]
          },
          "target": {
            "type": "string"
          }
        }
      },
      "PolicyStep": {
        "type": "object",
        "required": [
          "channel",
          "target",
          "wait_minutes"
        ],
        "properties": {
          "channel": {
            "type": "string",
            "enum": [
              "telegram",
              "email",
              "webhook",
              "call",
              "log"
            ]
          },
          "target": {
            "type": "string"
          },
          "wait_minutes": {
            "type": "integer"
          }
        },
        "description": "One notify/wait step of a personal notification policy (BETA-031)."
      },
      "NotificationPolicies": {
        "type": "object",
        "properties": {
          "default": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyStep"
            }
          },
          "important": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyStep"
            }
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "description": "A person: both an on-call roster entry and (optionally) an auth principal. Every key here is written unconditionally by CreateUser, so all are required except the opt-in policies.",
        "required": [
          "id",
          "name",
          "username",
          "email",
          "phone",
          "telegram_id",
          "slack_id",
          "mattermost_id",
          "timezone",
          "locale",
          "on_duty",
          "priority",
          "notification_targets",
          "role",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "telegram_id": {
            "type": "string"
          },
          "slack_id": {
            "type": "string",
            "description": "Slack account id. Attributes a Slack command or button tap to this person rather than to the bot; empty means the tap runs as the platform service principal."
          },
          "mattermost_id": {
            "type": "string",
            "description": "Mattermost account id. Same role as slack_id, for the Mattermost bot."
          },
          "timezone": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "enum": [
              "",
              "ru-RU",
              "en-US"
            ],
            "description": "Chosen UI language; empty means the UI follows the browser."
          },
          "on_duty": {
            "type": "boolean"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ]
          },
          "notification_targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationTarget"
            }
          },
          "notification_policies": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/NotificationPolicies"
              },
              {
                "type": "null"
              }
            ],
            "description": "Absent or null means the legacy path: every target fired at once."
          },
          "role": {
            "type": "string",
            "enum": [
              "",
              "viewer",
              "responder",
              "editor",
              "admin"
            ],
            "description": "Empty means a roster entry that cannot sign in — the default."
          },
          "provisioned_by": {
            "type": "string",
            "description": "Name of the infrastructure-as-code tool that created this object (\"terraform\"), absent when a person did. Present means the object is owned by that tool: the API refuses updates and deletes from anyone else, and the web UI disables its own edit controls, because an edit made here would be reverted by the next apply with nothing on screen having said so."
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "UserPage": {
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total_estimated": {
            "type": "boolean",
            "description": "Present and true when total is the database planner's estimate rather than an exact count: listings count exactly up to 10 000 rows."
          },
          "total_lower_bound": {
            "type": "boolean",
            "description": "Present and true when the exact count passed 10 000 and the planner could not estimate beyond that: total is then 10 001, a lower bound (\"10 000+\")."
          }
        }
      },
...
    },
    "parameters": {
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer"
        }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer"
        }
      },
      "Sort": {
        "name": "sort",
        "in": "query",
        "required": false,
        "description": "Column to order by, validated against the collection's own columns; an unknown name is a 400 rather than a silently ignored parameter. Defaults to the newest-first column where the collection has one (alert groups: last_received_at, alerts: received_at, notifications: created_at) and to id elsewhere.",
        "schema": {
          "type": "string"
        }
      },
      "Order": {
        "name": "order",
        "in": "query",
        "required": false,
        "description": "Sort direction. Applies to the requested sort column, or to the collection default when none is named.",
        "schema": {
          "type": "string",
          "enum": [
            "asc",
            "desc"
          ]
        }
      },
...
    },
    "requestBodies": {
      "Object": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Entity"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Error": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "User": {
        "description": "A user",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "UserList": {
        "description": "A paginated list of users",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserPage"
            }
          }
        }
      },
...
    }
  }
}